Implemented add/delete contacts and used args/form instead of direct uri
Implemented the add contact and delete contact in the contacts route. Also implemented the appropriate Handlers and DAOs, functional in the back end, then hardcoded a response into the handlers. I also made some slight modifications to the previous login route to use request.form instead of passing the data through the url. Finally, created a small list of keys to check that the keys passed through the forms/args are there.